requirements.txt
cloud-volume~=3.1
inquirer~=2.6
ipython
navis>=0.5.1
numpy
pandas
PyMCubes
pyperclip
requests
requests-futures
skeletor
trimesh
tqdm
python-catmaid>=2.0.2
